backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
httpx
python-dotenv
langchain
openai
